// protheus checker/front/src/app/models/verification.model.ts
export class Verification {
    id?:                any;
    user_id?:           any;
    email?:             string;
    cnpj?:              string;
    corporate_name?:    string;
    query_sql?:         string;
    frequency?:         string;
    last_verification?: string;
    next_verification?: string;
    created_at?:        string;
    updated_at?:        string;
    description?:       string;
}